actor-critic/requirements.txt
numpy>=1.21.0
torch>=1.10.0
gym>=0.21.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
tqdm>=4.62.0